chore: refactored existing code related to tokens and metadata
