+++ app/src/bcsc-theme/api/hooks/useTokens.tsx
@@ -2,7 +2,7 @@ import { useCallback, useMemo } from 'react'
 import { getDeviceCodeRequestBody } from 'react-native-bcsc-core'
 import BCSCApiClient from '../client'
 import { withAccount } from './withAccountGuard'
-import { BCSCAccountJWT, getBCSCAccountJWT } from '@/bcsc-theme/utils/bcsc-account-jwt'
+import { BCSCAccountJWT, getBCSCAccountJWT } from '@/bcsc-theme/utils/bcsc-account'
 
 export interface TokenStatusResponseData {
   access_token: string
@@ -17,14 +17,10 @@ export interface TokenStatusWithAccount extends TokenStatusResponseData {
   account: BCSCAccountJWT
 }
 
-export interface TokenStatusResponseDataWithDeviceCount extends TokenStatusResponseData {
-  bcsc_devices_count?: number
-}
-
 const useTokenApi = (apiClient: BCSCApiClient) => {
   const checkDeviceCodeStatus = useCallback(
     async (deviceCode: string, confirmationCode: string) => {
-      return withAccount<TokenStatusResponseDataWithDeviceCount>(async (account) => {
+      return withAccount<TokenStatusWithAccount>(async (account) => {
         const body = await getDeviceCodeRequestBody(deviceCode, account.clientID, account.issuer, confirmationCode)
 
         const { data } = await apiClient.post<TokenStatusResponseData>(apiClient.endpoints.token, body, {
@@ -36,7 +32,7 @@ const useTokenApi = (apiClient: BCSCApiClient) => {
 
         apiClient.tokens = { ...data, account: accountJwt }
 
-        return { ...data, bcsc_devices_count: accountJwt.bcsc_devices_count }
+        return apiClient.tokens
       })
     },
     [apiClient]
