+++ app/src/bcsc-theme/types/cards.ts
@@ -1,8 +1,8 @@
 export enum BCSCCardType {
-  Combined = 'BC Services Card Combo', // IAS value
-  Photo = 'BC Services Card Photo', // IAS value
-  NonPhoto = 'BC Services Card Non-Photo', // IAS value
-  Other = 'Non BC Services Card(s)', // Transform undefined card_type to 'Other' (ie: non-BCSC card)
+  Combined = 'BC Services Card Combo', // IAS compatible value
+  Photo = 'BC Services Card Photo', // IAS compatible value
+  NonPhoto = 'BC Services Card Non-Photo', // IAS compatible value
+  Other = 'Non BC Services Card(s)', // Local value for non-BCSC cards (IAS value is undefined)
   None = 'None', // Local value when no card is present
 }
